+++ traffic_portal/test/end_to_end/deliveryServices/delivery-services-spec.js
@@ -114,7 +114,7 @@ describe('Traffic Portal Delivery Services Suite', function() {
 		first.click();
 		expect(first.isSelected()).toBe(false);
 		let tableColumns = element.all(by.css('#deliveryServicesTable tr:first-child td'));
-		expect(tableColumns.count()).toBe(10);
+		expect(tableColumns.count()).toBe(11);
 	});
 
 	it('should update the ANY_MAP delivery service', function() {